scraper/requirements.txt
selenium>=4.18.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
TikTokApi>=6.2.0
playwright>=1.42.0
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
tenacity>=8.2.0
PyYAML>=6.0.0

